requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
requests==2.31.0
python-dotenv==1.0.0
transformers>=4.37.0
torch>=2.1.0
accelerate>=0.25.0
